skills/ai-data-engineering/examples/llamaindex-agents/requirements.txt
# LlamaIndex Agent Example Dependencies

# Core LlamaIndex
llama-index==0.12.6
llama-index-core==0.12.6

# LLM Providers
llama-index-llms-openai==0.3.18
llama-index-embeddings-openai==0.3.7
llama-index-embeddings-voyageai==0.3.2

# Vector Stores
llama-index-vector-stores-qdrant==0.5.1
qdrant-client==1.12.1

# Utilities
python-dotenv==1.0.1
